+++ examples/02_b_chained_example.rb
@@ -1,29 +1,13 @@
-class RakeCommander::Custom::Chained < RakeCommander
-  namespace :examples
-
+class RakeCommander::Custom::Chained < RakeCommander::Custom::Chainer
+  #namespace :examples # <-- inherited
   desc 'A task you want to chain to'
   task :chained
 
-  option '-s', '--say [SOMETHING]', "It says 'something'", default: %q(I don't know what to "say"...)
-  option '-b', '--debug', TrueClass, 'Whether to add additional context information to messages'
+  option_reopen :s, "It says 'something'", default: %q(I don't know what to "say"...)
+  option_remove :c, :w, :method
 
   def task(*_args)
     puts "Called !!"
     puts options[:s] if options[:s]
-  end
-
-  private
-
-  def puts(str)
-    return super unless options[:b]
-    super "#{app_id}   #{str}   #{thread_id}"
-  end
-
-  def app_id
-    "(#{self.class.task_fullname}: #{Rake.application.object_id})"
-  end
-
-  def thread_id
-    "(PID: #{Process.pid} ++ Thread: #{Thread.current.object_id})"
   end
 end