+++ lib/rake-commander/base/object_helpers.rb
@@ -1,0 +1,42 @@
+class RakeCommander
+  module Base
+    module ObjectHelpers
+      private
+
+      # Custom Object#deep_dup for rake commander
+      def commander_deep_dup(value, dup_objects: true, &dup_block)
+        case value
+        when Hash
+          commander_hash_deep_dup(value, dup_objects: dup_objects, &dup_block)
+        when Array
+          value.map {|v| commander_deep_dup(v, dup_objects: dup_objects, &dup_block)}
+        else
+          commander_object_deep_dup(value, dup_objects: true, &dup_block)
+        end
+      end
+
+      # Does the copy of the final object
+      def commander_object_deep_dup(value, dup_objects: true)
+        return yield(value)   if block_given?
+        return value          unless dup_objects
+        return value.deep_dup if value.respond_to?(:deep_dup)
+        value.dup
+      end
+
+      # Custom Hash#deep_dup for rake commander
+      def commander_hash_deep_dup(original, dup_objects: true, &dup_block)
+        raise ArgumentError, "Expecting Hash. Given: #{original.class}" unless original.is_a?(Hash)
+        hash = original.dup
+        original.each_pair do |key, value|
+          unless key.frozen? && key.is_a?(::String)
+            hash.delete(key)
+            key = commander_deep_dup(key, dup_objects: dup_objects, &dup_block) if dup_objects
+          end
+          value     = dup_objects ? commander_deep_dup(value, dup_objects: dup_objects, &dup_block) : value
+          hash[key] = value
+        end
+        hash
+      end
+    end
+  end
+end