+++ lib/rake-commander/options.rb
@@ -1,4 +1,5 @@
 require_relative 'options/name'
+require_relative 'options/description'
 require_relative 'options/arguments'
 require_relative 'options/result'
 require_relative 'options/error'
@@ -10,8 +11,9 @@ class RakeCommander
       def included(base)
         super(base)
         base.extend RakeCommander::Base::ClassHelpers
+        base.extend RakeCommander::Base::ClassInheritable
         base.extend ClassMethods
-        base.inheritable_attrs :banner, :options_hash
+        base.attr_inheritable :banner, :options_hash
         base.class_resolver :option_class, RakeCommander::Option
         base.send :include, RakeCommander::Options::Result
         base.send :include, RakeCommander::Options::Error
@@ -34,7 +36,6 @@ class RakeCommander
         return option_reopen(*args, override: override, **kargs, &block) if reopen
         opt = option_class.new(*args, **kargs, &block)
         add_to_options(opt, override: override)
-        self
       end
 
       # It re-opens an option for edition. If it does not exist, it **upserts** it.
@@ -45,17 +46,19 @@ class RakeCommander
       # @note when `short` and `name` are provided, `name` takes precedence over `short`
       #   in the lookup (to identify the existing option)
       def option_reopen(*args, override: false, **kargs, &block)
-        opt = option_class.new(*args, **kargs, sample: true, &block)
-        ref = options_hash.values_at(opt.name, opt.short).compact.first
-        return add_to_options(opt) unless ref
-        replace_in_options(ref, ref.merge(opt), override: override)
+        aux = option_class.new(*args, **kargs, sample: true, &block)
+        opt = options_hash.values_at(aux.name, aux.short).compact.first
+        return option(*args, **kargs, &block) unless opt
+        replace_in_options(opt, opt.merge(aux), override: override)
       end
 
-      # Explicitly installs the help of the options
-      # @note `OptionParser` already has `-h --help` as a native option.
-      def option_help
-        return false if options_hash.key?(:help) || options_hash.key?(:h)
-        option(:h, :help, 'Prints this help') { puts opts; exit(0) }
+      # Removes options with short or name `keys` from options
+      def option_remove(*keys)
+        keys.map do |key|
+          aux = option_class.new(key, sample: true)
+          opt = options_hash.values_at(aux.name, aux.short).compact.first
+          delete_from_options(opt) if opt
+        end
       end
 
       # Allows to use a set of options
@@ -70,16 +73,12 @@ class RakeCommander
         self
       end
 
-      # It parses the `ARGV`.
-      # @note it builds the `OptionParser` injecting the `middleware` block.
-      # @return [Hash] with `short` option as `key` and final value as `value`.
-
-      # Options
+      # It builds the `OptionParser` injecting the `middleware` block and parses `argv`
       # @note this method is extended in via the following modules:
       #   1. `RakeCommander::Options::Result`: makes the method to return a `Hash` with results,
       #     as well as captures/moves the **leftovers** to their own keyed argument.
-      #   2. `RakeCommander::Options:Error`:
-      # @param options_parser [Option::Parser] the options parser ready for parsing.
+      #   2. `RakeCommander::Options:Error`: adds error handling (i.e. forward to rake commander errors)
+      # @param argv [Array<String>] the command line arguments to be parsed.
       # @return [Array<String>] the **leftovers** of the `OptionParser#parse` call.
       def parse_options(argv = ARGV, &middleware)
         options_parser(&middleware).parse(argv)
@@ -109,7 +108,7 @@ class RakeCommander
       def options_parser(&middleware)
         new_options_parser do |opts|
           opts.banner = banner if banner
-          option_help
+          option_help(opts)
           free_shorts = implicit_shorts
 
           options.each do |opt|
@@ -119,6 +118,7 @@ class RakeCommander
         end
       end
 
+      # @return [OptionParser]
       def new_options_parser(&block)
         require 'optparse'
         OptionParser.new(&block)
@@ -127,6 +127,7 @@ class RakeCommander
       # The options indexed by the short and the name (so doubled up in the hash).
       # @param with_implicit [Boolean] whether free implicit shorts of declared options should be included
       #   among the keys (pointing to the specific option that has it implicit).
+      # @return [Hash] with Symbol name and shorts as keys, and `RakeCommander::Option` as values.
       def options_hash(with_implicit: false)
         @options_hash ||= {}
         return @options_hash unless with_implicit
@@ -150,17 +151,28 @@ class RakeCommander
 
       private
 
+      # Explicitly installs the help of the options
+      # @note `OptionParser` already has `-h --help` as a native option.
+      # @param opts [OptionParser] where the help will be added.
+      def option_help(opts)
+        return false if options_hash.key?(:help) || options_hash.key?(:h)
+        option(:h, :help, 'Prints this help') do
+          puts opts
+          exit(0)
+        end
+        self
+      end
+
       # @todo check that all the elements are of `option_class`
       # @return [Array<RakeCommander::Option>]
       def to_options(opts)
-        case
-        when opts.is_a?(Hash)
+        if opts.is_a?(Hash)
           opts.values
-        when opts.is_a?(RakeCommander::Options::Set)
+        elsif opts.is_a?(RakeCommander::Options::Set)
           opts.class.options
-        when opts <= RakeCommander::Options::Set
+        elsif opts <= RakeCommander::Options::Set
           opts.options
-        when opts.respond_to?(:to_a)
+        elsif opts.respond_to?(:to_a)
           opts.to_a
         end
       end
@@ -169,44 +181,55 @@ class RakeCommander
       # @todo add `exception` parameter, to trigger an exception
       #   when `opt` name or short are taken (and override is `false`)
       # @todo allow to specif if `:tail`, `:top` or `:base` (default)
-      # @return [Boolean] wheter or not it succeeded adding the option.
+      # @param opt [RakeCommander::Option] the option to be added.
+      # @param override [Boolean] whether we should continue, may this action override (an)other option(s).
+      # @return [RakeCommander::Option, NilClass] the option that was added, `nil` is returned otherwise.
       def add_to_options(opt, override: true)
         name_ref = respond_to?(:name)? " (#{name})" : ''
         if sprev = options_hash[opt.short]
-          return false unless override
+          return nil unless override
           puts "Warning#{name_ref}: Overriding option '#{sprev.name}' with short '#{sprev.short}' ('#{opt.name}')"
           delete_from_options(sprev)
         end
         if nprev = options_hash[opt.name]
-          return false unless override
+          return nil unless override
           puts "Warning#{name_ref}: Overriding option '#{nprev.short}' with name '#{nprev.name}' ('#{opt.short}')"
           delete_from_options(nprev)
         end
         options_hash[opt.name] = options_hash[opt.short] = opt
-        true
       end
 
       # Deletes an option from the `options_hash`
+      # @param opt [RakeCommander::Option] the option to be deleted.
+      # @return [RakeCommander::Option, NilClass] the option that was deleted, or `nil` otherwise.
       def delete_from_options(opt)
-        options_hash.delete(opt.short)
-        options_hash.delete(opt.name)
+        res = options_hash.delete(opt.short)
+        options_hash.delete(opt.name) || res
       end
 
-      # @return [Boolean] whether it succeeded replacing option `ref` with `opt`
+      # Replaces option `ref` with option `opt`.
+      # @note this method was added aiming to keep the same position for an option we override.
+      # @param ref [RakeCommander::Option] the option to be replaced.
+      # @param opt [RakeCommander::Option] the option that will replace `ref`.
+      # @param override [Boolean] whether we should continue, may there be collateral override to other options.
+      # @return [RakeCommander::Option, NilClass] `opt` if it was added, or `nil` otherwise.
       def replace_in_options(ref, opt, override: false)
         # Try to keep the same potition
         options_hash[ref.short] = options_hash[ref.name] = nil
-        add_to_options(opt, override: override).tap do |success|
-          next options_hash[ref.short] = options_hash[ref.name] = ref unless success
+        add_to_options(opt, override: override).tap do |added_opt|
+          # restore previous status
+          next options_hash[ref.short] = options_hash[ref.name] = ref unless added_opt
           delete_empty_keys(options_hash)
         end
       end
 
-      # Deletes all keys with `nil` as value
-      def delete_empty_keys(h)
-        h.dup.each do |k, v|
-          next unless v.nil?
-          h.delete(k)
+      # Deletes all keys that have `nil` as value
+      def delete_empty_keys(hash)
+        hash.tap do |_h|
+          hash.dup.each do |k, v|
+            next unless v.nil?
+            hash.delete(k)
+          end
         end
       end
     end