+++ lib/rake-commander/options/error/missing_argument.rb
@@ -1,9 +1,9 @@
 class RakeCommander
   module Options
     module Error
-      # Relates to options with missing required argument (when there's no `default` value) 
+      # Relates to options with missing required argument (when there's no `default` value)
       class MissingArgument < RakeCommander::Options::Error::Rely
-        option_regex /missing(?: required|) argument: (?<option>.+)/i.freeze
+        option_regex(/missing(?: required|) argument: (?<option>.+)/i.freeze)
       end
     end
   end